Points 1,2 and 4

+++ src/com/abir/photolotto/ImageLoader.java
@@ -23,19 +23,22 @@ import android.widget.ImageView;
 
 public class ImageLoader {
 
-	MemoryCache						memoryCache	= new MemoryCache();
-	FileCache						fileCache;
-	private Map<ImageView, String>	imageViews	= Collections.synchronizedMap(new WeakHashMap<ImageView, String>());
-	ExecutorService					executorService;
-	Handler							handler		= new Handler(); // handler to display images in UI thread
+	MemoryCache memoryCache = new MemoryCache();
+	FileCache fileCache;
+	private Map<ImageView, String> imageViews = Collections
+			.synchronizedMap(new WeakHashMap<ImageView, String>());
+	ExecutorService executorService;
+	Handler handler = new Handler(); // handler to display images in UI thread
+	Context context;
 
 	public ImageLoader(Context context) {
 		fileCache = new FileCache(context);
 		executorService = Executors.newFixedThreadPool(5);
+		this.context = context;
 	}
 
-	final int	stub_id	= R.drawable.black_thumb_image;
-	
+	final int stub_id = R.drawable.black_thumb_image;
+
 	public void displayImage(String url, ImageView imageView) {
 		imageViews.put(imageView, url);
 		Bitmap bitmap = memoryCache.get(url);
@@ -53,33 +56,44 @@ public class ImageLoader {
 	}
 
 	public Bitmap getBitmap(String url) {
-		File f = fileCache.getFile(url);
+		// if (Utils.isOnline(context)) {
+		String url1 = url.replaceAll(" ", "%20");
+		File file = new File(Constants.DIRECTORY_PATH, url1.substring(url1
+				.lastIndexOf("/") + 1));
+		if (file.exists() && !Utils.isOnline(context)) {
+			return decodeFile(file);
+		} else {
+			File f = fileCache.getFile(url);
 
-		// from SD cache
-		Bitmap b = decodeFile(f);
-		if (b != null)
-			return b;
+			// from SD cache
+			Bitmap b = decodeFile(f);
+			if (b != null)
+				return b;
 
-		// from web
-		try {
-			Bitmap bitmap = null;
-			URL imageUrl = new URL(url);
-			HttpURLConnection conn = (HttpURLConnection) imageUrl.openConnection();
-			conn.setConnectTimeout(30000);
-			conn.setReadTimeout(30000);
-			conn.setInstanceFollowRedirects(true);
-			InputStream is = conn.getInputStream();
-			OutputStream os = new FileOutputStream(f);
-			Utils.copyStream(is, os);
-			os.close();
-			bitmap = decodeFile(f);
-			return bitmap;
-		} catch (Throwable ex) {
-			ex.printStackTrace();
-			if (ex instanceof OutOfMemoryError)
-				memoryCache.clear();
-			return null;
+			// from web
+
+			try {
+				Bitmap bitmap = null;
+				URL imageUrl = new URL(url);
+				HttpURLConnection conn = (HttpURLConnection) imageUrl
+						.openConnection();
+				conn.setConnectTimeout(30000);
+				conn.setReadTimeout(30000);
+				conn.setInstanceFollowRedirects(true);
+				InputStream is = conn.getInputStream();
+				OutputStream os = new FileOutputStream(f);
+				Utils.copyStream(is, os);
+				os.close();
+				bitmap = decodeFile(f);
+				return bitmap;
+			} catch (Throwable ex) {
+				ex.printStackTrace();
+				if (ex instanceof OutOfMemoryError)
+					memoryCache.clear();
+				return null;
+			}
 		}
+
 	}
 
 	// decodes image and scales it to reduce memory consumption
@@ -87,30 +101,22 @@ public class ImageLoader {
 		try {
 			// decode image size
 			BitmapFactory.Options o = new BitmapFactory.Options();
-			//o.inJustDecodeBounds = true;
+			// o.inJustDecodeBounds = true;
 			FileInputStream stream1 = new FileInputStream(f);
 			Bitmap bitmap = BitmapFactory.decodeStream(stream1, null, o);
 			stream1.close();
 			// Find the correct scale value. It should be the power of 2.
 			/*
-			final int REQUIRED_SIZE = 70;
-			int width_tmp = o.outWidth, height_tmp = o.outHeight;
-			int scale = 1;
-			while (true) {
-				if (width_tmp / 2 < REQUIRED_SIZE || height_tmp / 2 < REQUIRED_SIZE)
-					break;
-				width_tmp /= 2;
-				height_tmp /= 2;
-				scale *= 2;
-			}
-
-			// decode with inSampleSize
-			BitmapFactory.Options o2 = new BitmapFactory.Options();
-			o2.inSampleSize = scale;
-			FileInputStream stream2 = new FileInputStream(f);
-			Bitmap bitmap = BitmapFactory.decodeStream(stream2, null, o2);
-			stream2.close();
-			*/
+			 * final int REQUIRED_SIZE = 70; int width_tmp = o.outWidth,
+			 * height_tmp = o.outHeight; int scale = 1; while (true) { if
+			 * (width_tmp / 2 < REQUIRED_SIZE || height_tmp / 2 < REQUIRED_SIZE)
+			 * break; width_tmp /= 2; height_tmp /= 2; scale *= 2; }
+			 * 
+			 * // decode with inSampleSize BitmapFactory.Options o2 = new
+			 * BitmapFactory.Options(); o2.inSampleSize = scale; FileInputStream
+			 * stream2 = new FileInputStream(f); Bitmap bitmap =
+			 * BitmapFactory.decodeStream(stream2, null, o2); stream2.close();
+			 */
 			return bitmap;
 		} catch (FileNotFoundException e) {
 		} catch (IOException e) {
@@ -121,8 +127,9 @@ public class ImageLoader {
 
 	// Task for the queue
 	private class PhotoToLoad {
-		public String		url;
-		public ImageView	imageView;
+		public String url;
+		public ImageView imageView;
+
 		public PhotoToLoad(String u, ImageView i) {
 			url = u;
 			imageView = i;
@@ -130,7 +137,8 @@ public class ImageLoader {
 	}
 
 	class PhotosLoader implements Runnable {
-		PhotoToLoad	photoToLoad;
+		PhotoToLoad photoToLoad;
+
 		PhotosLoader(PhotoToLoad photoToLoad) {
 			this.photoToLoad = photoToLoad;
 		}
@@ -161,12 +169,14 @@ public class ImageLoader {
 
 	// Used to display bitmap in the UI thread
 	class BitmapDisplayer implements Runnable {
-		Bitmap		bitmap;
-		PhotoToLoad	photoToLoad;
+		Bitmap bitmap;
+		PhotoToLoad photoToLoad;
+
 		public BitmapDisplayer(Bitmap b, PhotoToLoad p) {
 			bitmap = b;
 			photoToLoad = p;
 		}
+
 		public void run() {
 			if (imageViewReused(photoToLoad))
 				return;
